add start train function

+++ dotnet_ts_support/Models/TrainSetting.cs
@@ -17,29 +17,29 @@ namespace dotnet_ts_support.Models
         public int width { get; init; }
         public int height { get; init; }
         public int channels { get; init; }
-        public int baseLearningRate { get; init; }
-        public int gamma { get; init; }
+        public double baseLearningRate { get; init; }
+        public double gamma { get; init; }
         public int stepCount { get; init; }
         public int maxIteration { get; init; }
 
         public bool mirror { get; init; }
         public bool flip { get; init; }
         public bool rotation90 { get; init; }
-        public int zoom { get; init; }
-        public int tilt { get; init; }
-        public int shift { get; init; }
-        public int rotation { get; init; }
-        public int contrast { get; init; }
-        public int brightness { get; init; }
-        public int smoothFiltering { get; init; }
-        public int noise { get; init; }
-        public int colorNoise { get; init; }
-        public int partialFocus { get; init; }
-        public int shade { get; init; }
-        public int hue { get; init; }
-        public int saturation { get; init; }
+        public double zoom { get; init; }
+        public double tilt { get; init; }
+        public double shift { get; init; }
+        public double rotation { get; init; }
+        public double contrast { get; init; }
+        public double brightness { get; init; }
+        public double smoothFiltering { get; init; }
+        public double noise { get; init; }
+        public double colorNoise { get; init; }
+        public double partialFocus { get; init; }
+        public double shade { get; init; }
+        public double hue { get; init; }
+        public double saturation { get; init; }
         public int maxRandomAugmentCount { get; init; }
-        public int probability { get; init; }
+        public double probability { get; init; }
         public int borderMode { get; init; }
         
         public TrainSetting() { }
